Moved indexermgr to protected
